Initial AzureRM virtual machine resources
Because of the many additional resources required to properly configure
a vm. We will need to evaluate these constructs to abstract them the
most effective way across providers.

The following are required for an AzureRM virtual machine:
- azurerm_resource_group
- azurerm_virtual_network
- azurerm_subnet
- azurerm_network_interface

vSphere will also require more abstraction based on what we find with
AzureRM resources.