Moved VMs from underneath projects
Because DigitalOcean does not require anything to be in place to start
creating VMs, we have moved VMs out to their own structure under
resources. This also affected how firewalls, etc. are applied to VMs.
So, some additional tweaks were required to accomplish this.

There is still an issue with tags between modules if a VM moves. Will
address this in another issue.

Closes #27