Fixed DigitalOcean tags
DigitalOcean tags are now appropriately added across modules. For every
tag defined, a respective tag will be created for each environment. And
then for each droplet, etc. which can have tags applied will also have a
tag added for the environment as well. This will allow tagging to not
only work for just the defined tag, but also for the tag-environment as
well.

Closes #30